requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-RESTful==0.3.10

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Security
bcrypt==4.1.1
python-dotenv==1.0.0
WTForms==3.1.1

# File Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# PDF Generation
reportlab==4.0.7
Pillow==10.1.0

# Task Scheduling
APScheduler==3.10.4

# Email
email-validator==2.1.0

# HTTP and WSGI
Werkzeug==3.0.1
gunicorn==21.2.0
waitress==2.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
